test(core): collapse the model oracle cluster into two files
Merge the seven per-model oracle files into test_models_parametric.py
(Python model vs external oracle) and test_models_yaml.py (catalog
entry vs reference), with the shared builders in model_oracles.py.
Every node survives except the two default-catalog-entry duplicates
already covered by test_every_catalog_entry_builds.
